Clean up and minor refactor
