+++ pkg/integration_test.go
@@ -6,13 +6,7 @@ import (
 	"gopkg.in/yaml.v2"
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/chart/loader"
-	"helm.sh/helm/v3/pkg/chartutil"
-	kubefake "helm.sh/helm/v3/pkg/kube/fake"
-	"helm.sh/helm/v3/pkg/registry"
 	"helm.sh/helm/v3/pkg/release"
-	"helm.sh/helm/v3/pkg/storage"
-	"helm.sh/helm/v3/pkg/storage/driver"
-	"log"
 	"testing"
 )
 
@@ -24,7 +18,7 @@ func TestMigrator_IntegrationTests(t *testing.T) {
 agent:
   targetEnvironments: ["Development", "Test", "Prod"]`
 
-	chartV1Path := "test-charts/v1/"
+	chartV1Path := "test-charts/v1/" // Note that these aren't necessarily valid charts, they're just what we need for testing.
 	chartV2Path := "test-charts/v2/"
 
 	migration := `agent:
@@ -42,6 +36,7 @@ myKey: null
       - Prod
 `
 
+	// Load v1 chart
 	chV1, err := loader.Load(chartV1Path)
 	req.NoError(err, "Error loading chart v1")
 
@@ -50,13 +45,14 @@ myKey: null
 	install.Namespace = "default"
 	install.ReleaseName = "release-1"
 
-	cfgMap, err := YamlUnmarshal(customValuesV1)
+	// Install the v1 chart with the custom values
+	cfgMap, err := yamlUnmarshal(customValuesV1)
 	req.NoError(err, "Error unmarshalling custom values")
 
 	rel1, err := install.Run(chV1, cfgMap)
 	req.NoError(err, "Error installing chart v1")
 
-	//Make a migration
+	// Make a migration
 	ms := &MemoryMigrationSource{}
 	ms.AddMigrationData(2, migration)
 
@@ -64,25 +60,21 @@ myKey: null
 	migratedValues, err := Migrate(rel1.Config, nil, ms)
 	req.NoError(err, "Error migrating values")
 
-	// Create a new chart using the default values v2 and template v2
-	// apply migrated values to the new chart
+	// Load the v2 chart
 	chV2, err := loader.Load(chartV2Path)
 	req.NoError(err, "Error loading chart v2")
 
 	upAction := action.NewUpgrade(config)
 	upAction.ResetThenReuseValues = true
 
+	// Apply migrated values to the new chart
 	rel2, err := upAction.Run(rel1.Name, chV2, migratedValues)
 	req.NoError(err, "Error upgrading chart")
 	req.NotNil(rel2, "Release was not created")
 
 	// Now make sure it is actually upgraded
-	releases, err := config.Releases.ListReleases()
-	log.Println(releases[0].Name)
-	log.Println(releases[0].Version)
 	updatedRel, err := config.Releases.Get(rel2.Name, 2)
 	req.NoError(err, "Error getting updated release")
-
 	req.NotNil(updatedRel, "Updated Release is nil")
 
 	is.Equal(release.StatusDeployed, updatedRel.Info.Status)
@@ -90,44 +82,4 @@ myKey: null
 	updatedVals, _ := yaml.Marshal(updatedRel.Config)
 
 	is.Equal(expected, string(updatedVals))
-}
-
-func YamlUnmarshal(customValuesV1 string) (map[string]interface{}, error) {
-	var cfgMap map[string]interface{}
-	err := yaml.Unmarshal([]byte(customValuesV1), &cfgMap)
-	if err != nil {
-		return nil, err
-	}
-	return cfgMap, nil
-}
-
-// from https://github.com/helm/helm/blob/main/pkg/action/action_test.go#L39
-func actionConfigFixture(t *testing.T) *action.Configuration {
-	t.Helper()
-
-	registryClient, err := registry.NewClient()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return &action.Configuration{
-		Releases:       storage.Init(driver.NewMemory()),
-		KubeClient:     &kubefake.PrintingKubeClient{Out: &LogWriter{}},
-		Capabilities:   chartutil.DefaultCapabilities,
-		RegistryClient: registryClient,
-		Log: func(format string, v ...interface{}) {
-			t.Helper()
-			//if *verbose {
-			t.Logf(format, v...)
-			//}
-		},
-	}
-}
-
-type LogWriter struct {
-}
-
-func (*LogWriter) Write(p []byte) (n int, err error) {
-	log.Print(string(p))
-	return len(p), nil
 }
